[Nexters/covid-letter-web#47] 🎨 편지목록 empty case 컴포넌트 분리
