+++ src/pages/covid/letter/index.tsx
@@ -9,9 +9,8 @@ import {useState} from 'react'
 import {convertCommonDateFormat} from '$utils/date'
 import tw from 'twin.macro'
 import {FontOhsquare, FontOhsquareAir} from '$styles/utils/font'
-import {FlexCenter, FlexStart} from '$styles/utils/layout'
-import {useRouter} from 'next/router'
-import ROUTES from '$constants/routes'
+import {FlexStart} from '$styles/utils/layout'
+import EmptyLetterListContainer from '$components/letter/EmptyLetterListContainer'
 
 const Container = styled.div`
     ${tw`tw-bg-beige-300 tw-h-screen`}
@@ -37,32 +36,7 @@ const SubTitle = styled.div`
     ${tw`tw-text-left tw-text-base tw-text-primary-green-500`}
     margin-top: 0.8rem;
 `
-const EmptyListContainer = styled.div`
-    margin-top: 20.7rem;
-    
-    div {
-        ${FlexCenter}
-        span {
-            ${tw`tw-text-base tw-text-center tw-text-primary-green-500 tw-font-normal`}
-        }
-    }
-    
-    button {
-        margin-top: 3.2rem;
-        ${FlexCenter}
-        ${tw`tw-bg-primary-green-500 tw-w-full`}
-        height: 5.2rem;
-        border-radius: 0.4rem;
-        
-        span {
-            font-family: Cafe24 Ohsquare;
-            font-weight: bold;
-            font-size: 16px;
-            line-height: 25px;
-            ${tw`tw-text-grey-000`} 
-        }
-    }
-`
+
 const ListContainer = styled.ul``
 const ItemContainer = styled.li`
     margin: 1rem;
@@ -72,7 +46,6 @@ const ItemContainer = styled.li`
 
 const Letters = ({letters}: {letters: Letter[]}) => {
 
-    const router = useRouter()
     const [isShowEnvelope, setIsShowEnvelop] = useState(false)
     const [openedLetterId, setOpenedLetterId] = useState('')
 
@@ -83,10 +56,6 @@ const Letters = ({letters}: {letters: Letter[]}) => {
     const closeEnvelope = () => {
         setOpenedLetterId('')
         setIsShowEnvelop(false)
-    }
-
-    const createNewLetter = () => {
-        router.push({pathname: ROUTES.COVID.LETTER.OPTION})
     }
 
     const letterList = letters.length > 0
@@ -103,16 +72,7 @@ const Letters = ({letters}: {letters: Letter[]}) => {
                 ))}
             </ListContainer>
         )
-        : (
-            <EmptyListContainer>
-                <div>
-                    <span>작성된 편지가 없어.<br/>한번 편지를 쓰러 가볼까?</span>
-                </div>
-                <button onClick={createNewLetter}>
-                    <span>편지 작성</span>
-                </button>
-            </EmptyListContainer>
-        )
+        : <EmptyLetterListContainer />
 
     return (
         <>
